refactor: move controls to features level
