+++ src/features/game/index.ts
@@ -1,0 +1,2 @@
+export { default as RoundControls } from "./ui/controls/RoundControls";
+export { default as StageControls } from "./ui/controls/StageControls";
