+++ src/features/game/ui/controls/RoundControls.ts
@@ -1,8 +1,8 @@
-import { Component, label, option, select } from "../../../shared/ui";
-import { isValidSetting } from "../../../shared/helpers";
-import { Observer, Publisher } from "../../../entities/state";
+import { Component, label, option, select } from "../../../../shared/ui";
+import { isValidSetting } from "../../../../shared/helpers";
+import { Observer, Publisher } from "../../../../entities/state";
 
-import LevelsState from "../model/LevelsState";
+import LevelsState from "../../model/LevelsState";
 
 import styles from "./RoundControls.module.css";
