+++ src/features/game/ui/controls/StageControls.module.css
