+++ src/features/game/ui/controls/StageControls.ts
@@ -1,8 +1,8 @@
-import { Component, Button, Modal } from "../../../shared/ui";
+import { Component, Button, Modal } from "../../../../shared/ui";
 
-import type RoundState from "../model/RoundState";
+import type RoundState from "../../model/RoundState";
 
-import { Observer } from "../../../entities/state";
+import { Observer } from "../../../../entities/state";
 
 import styles from "./StageControls.module.css";
